ADD1/2: SciviewBridge.DisplayParamsProvider iface with classes:
Several content updating methods were requiring information from
opened BDV. Turned out that they always need only two pieces of
information -- this is now offered via the iface.

Two implementations to this iface are provided: One that wraps around
given BDV and fetches the required information from it -- this is the
original behaviour, and one implementation that repeats the
previously/lastly provided information -- and that does not need to have
a BDV window around.